Ensure the required config_file option is passed on the command line
Currently a path to the configuration file always needs to be set so
print an error message if it is not instead of trying to dereference a
null pointer when trying to load the configuration.

This also remove the test for whether the config file option is set
since that code is never reach now or before this change.